:sparkles:feat: 메인 게시판, 게시판 1개 목록, 게시글 검색 기능 구현 완료

+++ BE/deep/src/main/java/com/basic/deep/board/dto/BoardDetailResponseDTO.java
@@ -14,7 +14,7 @@ public class BoardDetailResponseDTO {
     private String memberRandom;
     private String memberFile;
     private LocalDateTime boardCreatedTime;
-    private int view;
+    private Long view;
     private List<String> img;
     private Boolean meLike;
     private Long like;
